feature(tags): support tag list, renaming, merging and removing
A new panel is introduced to the DatabaseMaintenance page which allows
 the user to list all registered tags in any tag entry with their internal
 ID and the number of items tagged with them.
Tags can now be renamed with the panel.
Tags can now be merged with the panel. This works by selecting multiple
 tags registered in an entry, then picking one from the selected tags.
 All items tagged with any of the tags selected will be re-tagged with
 the picked one. Deduplication is done automatically.
Tags can now be removed with the panel.
Unused tags can now be removed with a single button.
Updated dependencies.